requirements.txt
flask==2.0.1
lyricsgenius==3.1.1
python-dotenv==0.19.0
